Touchaction: Fix detection of tap when multi touch gestures are present
Fixes determination of if it is acceptable to prevent the browser default behavior of scrolling/zoom etc when the touch-action property is set to none and the browser does not provide native support for the touch-action property. It fails to differenciate between a tap gesture and a multi-touch gesture which results in click and hover events from not getting recognized when an element has multi-touch gesture associated with it.

Ref #838
Ref #808
Closes #762